requirements.txt
oauth2
python-dotenv
plurk-oauth
click
